Create Liquidity
This code gets the required data from Mixpanel to calculate liquidity for Real Estate Products. The final table produces the volume of liquid ads by day and vertical. It can then be linked with the table with the NNLs also produced in the code to calculate liquidity. NNLs are sourced from Silver and replies from Mixpanel.